Sample Model Using Class
This is a basic View that contains a text field that collects three
vehicle fields: year, make, and model.  The button Show Description is
pushed and the three fields are shown as one text label.

The Model file declares a Model class which uses private vars.  It
takes the input and calculates a variable combing all the fields.